Refactor ObsEdit header; move code into Header component
